14-12-2021

+++ services/userFolderService.js
@@ -83,7 +83,10 @@ function UserFolderService(userID){
 
     this.clearUserFolder = ()=>{
         try{
-           fs.rmdirSync(path.join(__dirname,"../files",this.userID), { recursive: true, force: true })
+            if(fs.existsSync(path.join(__dirname,"../files",this.userID))){
+                fs.rmdirSync(path.join(__dirname,"../files",this.userID), { recursive: true, force: true })
+            }
+      
         }
         catch(error){
             console.log(error)
